update kubebuilder and minor refactor
