+++ .vscode/launch.json
@@ -9,7 +9,7 @@
             "type": "go",
             "request": "launch",
             "mode": "auto",
-            "program": "${workspaceRoot}/cmd/dast-operator/main.go",
+            "program": "${workspaceRoot}/main.go",
             "env": {},
             "args": []
         }
