+++ Makefile
@@ -6,48 +6,46 @@ ANALYZER_IMG ?= ${ORGANIZATION}/dast-analyzer:latest
 # Produce CRDs that work back to Kubernetes 1.11 (no version conversion)
 CRD_OPTIONS ?= "crd:trivialVersions=true"
 
+# Get the currently used golang install path (in GOPATH/bin, unless GOBIN is set)
+ifeq (,$(shell go env GOBIN))
+GOBIN=$(shell go env GOPATH)/bin
+else
+GOBIN=$(shell go env GOBIN)
+endif
+
 all: manager
 
-
-# Install kustomize
-install-kustomize:
-	@ if ! which bin/kustomize &>/dev/null; then\
-		scripts/install_kustomize.sh;\
-	fi
-
-# Install kubebuilder
-install-kubebuilder:
-	@ if ! which bin/kubebuilder/bin/kubebuilder &>/dev/null; then\
-		scripts/install_kubebuilder.sh;\
-	fi
-
 # Run tests
-test: install-kubebuilder generate fmt vet manifests
-	KUBEBUILDER_ASSETS="$${PWD}/bin/kubebuilder/bin" go test ./api/... ./controllers/... ./pkg/... ./webhooks/... -coverprofile cover.out
+test: generate fmt vet manifests
+	go test ./... -coverprofile cover.out
 
 # Build manager binary
 manager: generate fmt vet
-	go build -o bin/dast-operator ./cmd/dast-operator/...
+	go build -o bin/manager main.go
 
 analyzer:
 	cd cmd/dynamic-analyzer; go build -o ../../bin/dynamic-analyzer ./... ;cd ../..
 
 # Run against the configured Kubernetes cluster in ~/.kube/config
-run: generate fmt vet
-	go run ./cmd/dast-operator/main.go
+run: generate fmt vet manifests
+	go run ./main.go
 
 # Install CRDs into a cluster
 install: manifests
-	kubectl apply -f config/crd/bases
+	kustomize build config/crd | kubectl apply -f -
+
+# Uninstall CRDs from a cluster
+uninstall: manifests
+	kustomize build config/crd | kubectl delete -f -
 
 # Deploy controller in the configured Kubernetes cluster in ~/.kube/config
-deploy: install-kustomize manifests
-	kubectl apply -f config/crd/bases
-	./bin/kustomize build config/default | kubectl apply -f -
+deploy: manifests
+	cd config/manager && kustomize edit set image controller=${IMG}
+	kustomize build config/default | kubectl apply -f -
 
 # Generate manifests e.g. CRD, RBAC etc.
 manifests: controller-gen
-	$(CONTROLLER_GEN) $(CRD_OPTIONS) rbac:roleName=manager-role webhook paths="./api/...;./controllers/..." output:crd:artifacts:config=config/crd/bases
+	$(CONTROLLER_GEN) $(CRD_OPTIONS) rbac:roleName=manager-role webhook paths="./..." output:crd:artifacts:config=config/crd/bases
 
 # Run go fmt against code
 fmt:
@@ -59,13 +57,11 @@ vet:
 
 # Generate code
 generate: controller-gen
-	$(CONTROLLER_GEN) object:headerFile=./hack/boilerplate.go.txt paths=./api/...
+	$(CONTROLLER_GEN) object:headerFile="hack/boilerplate.go.txt" paths="./..."
 
 # Build the docker image
 docker-build: test
 	docker build . -t ${IMG}
-	@echo "updating kustomize image patch file for manager resource"
-	sed -i'' -e 's@image: .*@image: '"${IMG}"'@' ./config/default/manager_image_patch.yaml
 
 docker-analyzer:
 	docker build . -t ${ANALYZER_IMG} -f Dockerfile-analyzer
@@ -78,8 +74,15 @@ docker-push:
 # download controller-gen if necessary
 controller-gen:
 ifeq (, $(shell which controller-gen))
-	go get sigs.k8s.io/controller-tools/cmd/controller-gen@v0.2.0-beta.1
-CONTROLLER_GEN=$(shell go env GOPATH)/bin/controller-gen
+	@{ \
+	set -e ;\
+	CONTROLLER_GEN_TMP_DIR=$$(mktemp -d) ;\
+	cd $$CONTROLLER_GEN_TMP_DIR ;\
+	go mod init tmp ;\
+	go get sigs.k8s.io/controller-tools/cmd/controller-gen@v0.2.5 ;\
+	rm -rf $$CONTROLLER_GEN_TMP_DIR ;\
+	}
+CONTROLLER_GEN=$(GOBIN)/controller-gen
 else
 CONTROLLER_GEN=$(shell which controller-gen)
 endif
