+++ PROJECT
@@ -1,3 +1,7 @@
-version: "2"
 domain: banzaicloud.io
 repo: github.com/banzaicloud/dast-operator
+resources:
+- group: security
+  kind: Dast
+  version: v1alpha1
+version: "2"
