+++ api/v1alpha1/zz_generated.deepcopy.go
@@ -76,7 +76,7 @@ func (in *Dast) DeepCopyObject() runtime.Object {
 func (in *DastList) DeepCopyInto(out *DastList) {
 	*out = *in
 	out.TypeMeta = in.TypeMeta
-	out.ListMeta = in.ListMeta
+	in.ListMeta.DeepCopyInto(&out.ListMeta)
 	if in.Items != nil {
 		in, out := &in.Items, &out.Items
 		*out = make([]Dast, len(*in))
