+++ config/crd/kustomization.yaml
@@ -3,12 +3,18 @@
 # It should be run by config/default
 resources:
 - bases/security.banzaicloud.io_dasts.yaml
-# +kubebuilder:scaffold:kustomizeresource
+# +kubebuilder:scaffold:crdkustomizeresource
 
-patches:
+patchesStrategicMerge:
+# [WEBHOOK] To enable webhook, uncomment all the sections with [WEBHOOK] prefix.
 # patches here are for enabling the conversion webhook for each CRD
 #- patches/webhook_in_dasts.yaml
-# +kubebuilder:scaffold:kustomizepatch
+# +kubebuilder:scaffold:crdkustomizewebhookpatch
+
+# [CERTMANAGER] To enable webhook, uncomment all the sections with [CERTMANAGER] prefix.
+# patches here are for enabling the CA injection for each CRD
+#- patches/cainjection_in_dasts.yaml
+# +kubebuilder:scaffold:crdkustomizecainjectionpatch
 
 # the following config is for teaching kustomize how to do kustomization for CRDs.
 configurations:
