--- config/crd/patches/webhook_in_dast.yaml
@@ -1,18 +1,0 @@
-# The following patch enables conversion webhook for CRDw
-apiVersion: apiextensions.k8s.io/v1beta1
-kind: CustomResourceDefinition
-metadata:
-  annotations:
-    certmanager.k8s.io/inject-ca-from: $(NAMESPACE)/$(CERTIFICATENAME)
-  name: dasts.security.banzaicloud.io
-spec:
-  conversion:
-    strategy: Webhook
-    webhookClientConfig:
-      # this is "\n" used as a placeholder, otherwise it will be rejected by the apiserver for being blank,
-      # but we're going to set it later using the cert-manager (or potentially a patch if not using cert-manager)
-      caBundle: Cg==
-      service:
-        namespace: $(NAMESPACE)
-        name: webhook-service
-        path: /convert-dast
