--- config/default/manager_image_patch.yaml-e
@@ -1,12 +1,0 @@
-apiVersion: apps/v1
-kind: StatefulSet
-metadata:
-  name: dast-operator
-  namespace: system
-spec:
-  template:
-    spec:
-      containers:
-      # Change the value of image field below to your controller image URL
-      - image: banzaicloud/dast-operator:latest
-        name: manager
