+++ config/manager/kustomization.yaml
@@ -1,3 +1,2 @@
 resources:
 - manager.yaml
-- webhook_config.yaml
