--- config/manager/webhook_config.yaml
@@ -1,20 +1,0 @@
-apiVersion: admissionregistration.k8s.io/v1beta1
-kind: ValidatingWebhookConfiguration
-metadata:
-  name: validation-webhook
-  labels:
-    app: admission-webhook
-webhooks:
-  - name: dast.security.banzaicloud.io
-    failurePolicy: Fail
-    clientConfig:
-      service:
-        name: webhook-service
-        namespace: system
-        path: "/ingress"
-      caBundle: LS0tLS1CRUdJTiBDRVJUSUZJQ0FURS0tLS0tCk1JSURiekNDQWxlZ0F3SUJBZ0lKQU5EYTRoNG1KYWl1TUEwR0NTcUdTSWIzRFFFQkN3VUFNRTR4Q3pBSkJnTlYKQkFZVEFsVlRNUk13RVFZRFZRUUlEQXBEWVd4cFptOXlibWxoTVJJd0VBWURWUVFIREFsVWFHVWdRMnh2ZFdReApGakFVQmdOVkJBb01EVTE1SUVOdmJYQmhibmtnUTBFd0hoY05NVGt3T1RFeU1UTTBPRFF3V2hjTk1qQXdNVEV3Ck1UTTBPRFF3V2pCT01Rc3dDUVlEVlFRR0V3SlZVekVUTUJFR0ExVUVDQXdLUTJGc2FXWnZjbTVwWVRFU01CQUcKQTFVRUJ3d0pWR2hsSUVOc2IzVmtNUll3RkFZRFZRUUtEQTFOZVNCRGIyMXdZVzU1SUVOQk1JSUJJakFOQmdrcQpoa2lHOXcwQkFRRUZBQU9DQVE4QU1JSUJDZ0tDQVFFQXRyQW0wQitUMENnekJvWWpCNzBtTEdjZFBOcVpMOVk2CndGd0h4NGNrYjBHbzRJSGwyaDRNemxnK2txejd4WjNZWVFzbU5vNjdrNURDZmNYQ05NMnN6WjR5bUZVN0x0SzEKZ2tjQ2ZwOXRmemc0ZVBHMW1TSU5hb0JNQUhZMFJxTkk2NFhueUtDVmlYTTBZZEl6UEtDeWRNK25jL3hJRHV1QwpDYXlvTnVLM0ZETmMyMmtFbFJCUEpmbTVUT3pwbkRqQlFHNTNkNXZDUEx2UmhBdlFON0xjTzl3YlkwdjFVLzA4CkRQNnI2UWNSdE1DNmprekJmbW5SM2lZQktBUm1TdVB0ZzFYdFg5Nk1yRHIzSWRib04yQVBPMDBwck5xak9aZloKL3VFMVBjRy92ZWUxVGpwVCtpNDdRUXE2TWx3WVVNbU9tZko1N01MaDYybGtJZjFPY2NoUE13SURBUUFCbzFBdwpUakFkQmdOVkhRNEVGZ1FVZmg3R3g4Zm42QlZ2L1RXcFg5SEk4UDdFditrd0h3WURWUjBqQkJnd0ZvQVVmaDdHCng4Zm42QlZ2L1RXcFg5SEk4UDdFditrd0RBWURWUjBUQkFVd0F3RUIvekFOQmdrcWhraUc5dzBCQVFzRkFBT0MKQVFFQUFXRDIrMkROYlFPMCtkcG5jNGNxN0VuNSs0RUIvYytINDRkRnpjMzJtS0p5dHhFV2R3VXREUCtxZ296OApObzdJNEFHdlU1bXFmdmdSSUN2TnVSdWpLdmZHaUhaRDFMNnBKVEhwQjFLWEtYaS80cnMzSkNuRld6V3ZhNHVkCkk5SGZVRUpDYTNueG9SUFJHK2Q4RnkweUdKdGlQN1ZKWTI5blZsbkMzcDZuM25LeE5UTXNxZ2dXa2tRYWEwbFAKWmUyeW5yMDNwL2lwemNqKzBTZXMrVGZhWlhsaTUxM1BsVU9hZmZMbExDR1pXOHQ4T00zblFFVlg0VzNOZDA4Mgo4d1o0WGJuNHlQU0RPbXFaYmlmQ2c0K05acU9yMFEyTXJ1RW1xZzdOTGREN2pXYk9GUVhWd0VaM0pQQVN4ZC95CmRDZms2NHNLOUZtMklzNEcwd3NFK0pBSVR3PT0KLS0tLS1FTkQgQ0VSVElGSUNBVEUtLS0tLQo=
-    rules:
-      - operations: [ "CREATE" ]
-        apiGroups: ["extensions"]
-        apiVersions: ["v1beta1"]
-        resources: ["ingresses"]
