+++ config/rbac/auth_proxy_service.yaml
@@ -1,13 +1,8 @@
 apiVersion: v1
 kind: Service
 metadata:
-  annotations:
-    prometheus.io/port: "8443"
-    prometheus.io/scheme: https
-    prometheus.io/scrape: "true"
   labels:
     control-plane: controller-manager
-    controller-tools.k8s.io: "1.0"
   name: controller-manager-metrics-service
   namespace: system
 spec:
@@ -17,4 +12,3 @@ spec:
     targetPort: https
   selector:
     control-plane: controller-manager
-    controller-tools.k8s.io: "1.0"
