+++ config/rbac/dast_editor_role.yaml
@@ -1,0 +1,24 @@
+# permissions for end users to edit dasts.
+apiVersion: rbac.authorization.k8s.io/v1
+kind: ClusterRole
+metadata:
+  name: dast-editor-role
+rules:
+- apiGroups:
+  - security.banzaicloud.io
+  resources:
+  - dasts
+  verbs:
+  - create
+  - delete
+  - get
+  - list
+  - patch
+  - update
+  - watch
+- apiGroups:
+  - security.banzaicloud.io
+  resources:
+  - dasts/status
+  verbs:
+  - get
