+++ config/rbac/dast_viewer_role.yaml
@@ -1,0 +1,20 @@
+# permissions for end users to view dasts.
+apiVersion: rbac.authorization.k8s.io/v1
+kind: ClusterRole
+metadata:
+  name: dast-viewer-role
+rules:
+- apiGroups:
+  - security.banzaicloud.io
+  resources:
+  - dasts
+  verbs:
+  - get
+  - list
+  - watch
+- apiGroups:
+  - security.banzaicloud.io
+  resources:
+  - dasts/status
+  verbs:
+  - get
