+++ config/rbac/kustomization.yaml
@@ -1,9 +1,12 @@
 resources:
 - role.yaml
 - role_binding.yaml
-# Comment the following 3 lines if you want to disable
+- leader_election_role.yaml
+- leader_election_role_binding.yaml
+# Comment the following 4 lines if you want to disable
 # the auth proxy (https://github.com/brancz/kube-rbac-proxy)
 # which protects your /metrics endpoint.
 - auth_proxy_service.yaml
 - auth_proxy_role.yaml
 - auth_proxy_role_binding.yaml
+- auth_proxy_client_clusterrole.yaml
