--- config/samples/security_v1alpha1_dast_external.yaml
@@ -1,12 +1,0 @@
-apiVersion: security.banzaicloud.io/v1alpha1
-kind: Dast
-metadata:
-  name: dast-sample-external
-spec:
-  zapproxy:
-    name: dast-test-external
-    apikey: abcd1234
-  analyzer:
-    image: banzaicloud/dast-analyzer:latest
-    name: external-test
-    target: https://example.com
