--- config/samples/test_ingress.yaml
@@ -1,17 +1,0 @@
-apiVersion: extensions/v1beta1
-kind: Ingress
-metadata:
-  name: test-ingress
-  annotations:
-    dast.security.banzaicloud.io/medium: "2"
-    dast.security.banzaicloud.io/low: "5"
-    dast.security.banzaicloud.io/informational: "10"
-    nginx.ingress.kubernetes.io/rewrite-target: /
-spec:
-  rules:
-  - http:
-      paths:
-      - path: /
-        backend:
-          serviceName: test-service
-          servicePort: 80
