--- config/samples/test_secvice.yaml
@@ -1,39 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: test-deployment
-  namespace: test
-  labels:
-    app: nginx
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: nginx
-      secscan: dast
-  template:
-    metadata:
-      labels:
-        app: nginx
-        secscan: dast
-    spec:
-      containers:
-      - name: nginx
-        image: nginx:1.16.0-alpine
-        ports:
-        - containerPort: 80
----
-apiVersion: v1
-kind: Service
-metadata:
-  name: test-service
-  annotations:
-    dast.security.banzaicloud.io/zapproxy: "dast-test"
-    dast.security.banzaicloud.io/zapproxy-namespace: "zapproxy"
-spec:
-  selector:
-    app: nginx
-    secscan: dast
-  ports:
-  - port: 80
-    targetPort: 80
