+++ config/webhook/kustomization.yaml
@@ -4,18 +4,3 @@ resources:
 
 configurations:
 - kustomizeconfig.yaml
-
-# the following config is for teaching kustomize how to do var substitution
-vars:
-- name: NAMESPACE
-  objref:
-    kind: Service
-    version: v1
-    name: webhook-service
-  fieldref:
-    fieldpath: metadata.namespace
-- name: SERVICENAME
-  objref:
-    kind: Service
-    version: v1
-    name: webhook-service
