+++ config/webhook/service.yaml
@@ -2,12 +2,11 @@
 apiVersion: v1
 kind: Service
 metadata:
-  name: dast-webhook-service
+  name: webhook-service
   namespace: system
 spec:
   ports:
     - port: 443
-      targetPort: 8443
+      targetPort: 9443
   selector:
-    control-plane: dast-operator
-    controller-tools.k8s.io: "1.0"
+    control-plane: controller-manager
